Organize src folder and remove some extra dependencies we didn’t end up using.
