+++ src/components/About.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import houstonhack from './houstonhack.jpg';
-import mountain from './mountain.png';
-import christian from './christian.png';
-import cameron from './cameron.png';
-import robyn from './robyn.png';
-import dylan from './dylan.png';
-import daniel from './daniel.png';
-import eric from './eric.png';
-import elaine from './elaine.png';
-import train from './logo3.png';
+import houstonhack from '../images/houstonhack.jpg';
+import mountain from '../images/mountain.png';
+import christian from '../images/christian.png';
+import cameron from '../images/cameron.png';
+import robyn from '../images/robyn.png';
+import dylan from '../images/dylan.png';
+import daniel from '../images/daniel.png';
+import eric from '../images/eric.png';
+import elaine from '../images/elaine.png';
+import logo from '../images/logo3-purple.png';
 
 class About extends React.Component {
   constructor(props) {
@@ -34,7 +34,7 @@ class About extends React.Component {
 		    	</div>
 		    	<div className="col-sm-6">
 			    	<h2>Houston Hackathon 2017</h2>
-			    	<p>This project was done as part of the 2017 Houston Hackathon,
+			    	<p>This project started as part of the 2017 Houston Hackathon,
 			    	where teams spend a weekend building applications that help solve
 			    	civic issues within the city of Houston.</p>
 		    	</div>
@@ -42,7 +42,7 @@ class About extends React.Component {
 		    <div className="row">
 		    	<div className="col-sm-6">
 			    	<h2>Mission</h2>
-			    	<p>There are two main public libraries in Houston; the Houston Public Library and Harris 
+			    	<p>There are two public library systems in Houston; the Houston Public Library and Harris
 			    	County Public Library. To search for digital books, people are currently forced to perform
 			    	two separate searches, one at each of the library's sites. Our goal is to provide a platform
 			    	that allows users to search for and check out digital books from both libraries. We are also 
@@ -99,7 +99,7 @@ class About extends React.Component {
                     <h2>Houston Book Link!</h2>
                 </div>
 				<div className="col-xs-12 col-sm-2 col-sm-offset-5">
-					<img className="media-object train" src={train} alt="Houston Book Link logo" width="48"/>
+					<img className="media-object train" src={logo} alt="Houston Book Link logo" width="48"/>
 				</div>
             </div>
 	    </div>
