+++ src/index.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
 import registerServiceWorker from './registerServiceWorker';
-import './index.css';
+import App from './components/App';
+
 import 'bootstrap/dist/css/bootstrap.css';
+import './index.css';
 
 
 ReactDOM.render(<App />, document.getElementById('root'));
